Corrected NHANES data and Juptyer notebook
Whereas poverty index previously could be up to 999, poverty index now is maximum of 9.98, in line with the NHANES definition of poverty index. Poverty index that was previously '999' is now 'Nan', since 999 is a code for 'unknown'. Whereas "Red blood cells" previously was maximum 88.8, red blood cells (in millions/mm^3) was scaled down by a factor of ten, to correspond correctly to the NHANES DU4800 documentation. 'Unacceptable data' and 'Blank, but applicable' (previously 77.7 and 88.8) are now 'NaN'.