Reduce memory usage for interaction values in big trees
The array for the unique features used in calculating the interaction values in tree_shap.h allocates max_nodes integers per tree to save the unique features in a tree. For big trees, this is too much, as the number of features is an upper bound for the amount of unique features.